+++ examples/resolved_annotations_example/lib/annotations2.dart
@@ -1,0 +1,4 @@
+class WrapGetter {
+  final String getterName;
+  const WrapGetter(this.getterName);
+}