+++ examples/resolved_annotations_example/lib/main.g.dart
@@ -1,0 +1,7 @@
+part of 'main.dart';
+
+extension MyClassExtension on MyClass {
+  String get data {
+    return _internal;
+  }
+}